fix: Don't panic if image decoding fails
Previous code used `unwrap` on the result from `decode`, which could
panic if `decode` returned an error.

Adjust this to return a Ruby encoding exception (which might be the
closest match).

Simplify the code a bit by using `map_err` to create the error and `?`
to return early in case of error, instead of using `match`.

While I'm here, add a little documentation for the function, and run
through `rustfmt`.